chore (ci): use elasticpath docker images in CI process
This change uses base images maintained by Elastic Path for the CI
process. The docker image is pulled from quay.io. Right now it is
using my personal token. When this project is officially setup, it
should be updated to use a more generic account with restricted
access. This change is also to illustrate that necessary CI secrets
can be managed in the GitLab mirror repo where CI is running.