fix: update steps that builds the server app
The current configuration of the project is able to start the
dev instance of the GraphQL server. However it fails to start
when it is built. The failure is mostly due to mixed use of
typescript and non-typescript files. For example: the typedef
files are '.graphql' extension but they are not copied to the
final distibution.

Applied necessary configurations to both tsconfig as well as
build config. Added instructions into Readme on how to build
the non-dev version of the server as well as suggestions on
several deployment options.